recovery_backup_2300/restore_tmp/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
google-genai
pydantic
python-multipart
aiofiles
requests
